tests/wpt/web-platform-tests/encrypted-media/Google/encrypted-media-session-closed-event.html

<!DOCTYPE html>
<!-- Copyright © 2016 Chromium authors and World Wide Web Consortium, (Massachusetts Institute of Technology, ERCIM, Keio University, Beihang). -->
<html>
    <head>
        <title>Test MediaKeySession closed event</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script>
            async_test(function(test)
            {
                var initDataType;
                var initData;
                var mediaKeySession;
                getSupportedInitDataType().then(function(type) {
                    initDataType = type;
                    initData = getInitData(initDataType);
                    return navigator.requestMediaKeySystemAccess('org.w3.clearkey', [{}]);
                }).then(function(access) {
                    return access.createMediaKeys();
                }).then(function(mediaKeys) {
                    mediaKeySession = mediaKeys.createSession();
                    return mediaKeySession.generateRequest(initDataType, initData);
                }).then(function() {
                    // Wait for the session to be closed.
                    mediaKeySession.closed.then(function(result) {
                        assert_equals(result, undefined);
                        // Now that the session is closed, verify that the
                        // closed attribute immediately returns a fulfilled
                        // promise.
                        return mediaKeySession.closed;
                    }).then(function(result) {
                        assert_equals(result, undefined);
                        test.done();
                    });

                    // release() should result in the closed promise being
                    // fulfilled.
                    return mediaKeySession.close();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'Test MediaKeySession closed event.');
        </script>
    </body>
</html>
